requirements.txt
pytesseract
pdf2image
Pillow
opencv-python
langdetect
pyinstaller
logging
numpy